% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPP_LUE_model.R
\name{addPredictor_tn}
\alias{addPredictor_tn}
\title{add previous time step GPP as a new variable}
\usage{
addPredictor_tn(x)
}
\arguments{
\item{x}{Data.table with the column of ydn and GPP}
}
\description{
add previous time step GPP as a new variable
}
