% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_radiation.R
\name{cal_Ra}
\alias{cal_Ra}
\title{Extraterrestrial radiation}
\usage{
cal_Ra(lat, doy)
}
\arguments{
\item{lat}{latitude}

\item{doy}{doy is the number of the day in the year between 1 (1 January)
and 365 or 366 (31 December)}
}
\value{
Ra, MJ m-2 d-1
}
\description{
Extraterrestrial radiation
}
