% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPP_LUE_model.R
\name{cal_Wscalar}
\alias{cal_Wscalar}
\title{cal_Wscalar}
\usage{
cal_Wscalar(d_mod09a1, pheno_T)
}
\arguments{
\item{d_mod09a1}{data.frame with the columns of \verb{site, date, year2, LSWI, QC_flag}}

\item{pheno_T}{data.frame with the columsn of \verb{site, year2, sos_date, eos_date}.}
}
\description{
cal_Wscalar
}
