% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_Ipaper.R
\name{check_file}
\alias{check_file}
\title{check_file}
\usage{
check_file(file, outdir)
}
\arguments{
\item{file}{}
}
\value{
A integer indicating file status
\itemize{
\item \code{0}  : file  already exist
\item \code{-1} : file isn't exist, and can't open it
\item \code{1}  : ok
}
}
\description{
check_file
}
\keyword{internal}
