% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_async.R
\name{check_sensitivity}
\alias{check_sensitivity}
\alias{check_sensitivity_async}
\title{check_sensitivity}
\usage{
check_sensitivity(x, predictors)

check_sensitivity_async(x, predictors)
}
\description{
check_sensitivity
}
\examples{
# predictors <- c("EVI", "Rs", "TA", "Prcp", "VPD", "APAR")#[-6]#[-c(1, 2)]
# check_sensitivity(x, predictors)
}
