% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_lag.R
\name{corr_lag}
\alias{corr_lag}
\alias{corr_lags}
\alias{find_x_lag}
\title{Lag correlation between two variables}
\usage{
corr_lag(x, y = x, index = NULL, lag = 1, dates, doy_lims, verbose = FALSE)

corr_lags(x, y, index, lags, dates, doy_lims, ...)

find_x_lag(x, index, lag)
}
\arguments{
\item{x, y}{A numeric vector}

\item{index}{A numeric vector, the subscript of x, e.g. \code{ydn}}

\item{lag}{The step of lag time}

\item{dates}{Dates vector, dates of x}

\item{doy_lims}{constrain index in the range of \code{period}}

\item{...}{other parameters to \code{\link[=corr_lag]{corr_lag()}}}
}
\value{
\itemize{
\item \code{cor}: pearson correlation of lag step
\item \code{lag}: positive (negative) value means x is ahead (behind) of y about \code{lag} steps.
}
}
\description{
Lag correlation between two variables
}
\examples{
# examples
x = sin(1:365)
y = x
index = 1:365
dates = seq(as.Date('2010-01-01'), as.Date('2010-12-31'), 'day')
corr_lags(x, y, index, dates, lag = 1)

}
