% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_reason.R
\name{define_period}
\alias{define_period}
\title{define_period
Define growing periods according to history phenology metrics.}
\usage{
define_period(his_phenology, delta = 8)
}
\description{
define_period
Define growing periods according to history phenology metrics.
}
