% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{df_North}
\alias{df_North}
\title{GPP observation in North Hemisphere

It has at least 70\\% valid observation in each site-year (116 sites in total).}
\description{
GPP observation in North Hemisphere

It has at least 70\\% valid observation in each site-year (116 sites in total).
}
