% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elasticity.R
\name{Elasticity_GPP}
\alias{Elasticity_GPP}
\title{Get 1th back forward derivate}
\usage{
Elasticity_GPP(x, predictors)
}
\arguments{
\item{x}{A data.table, at least with the column of varnames}
}
\description{
Get 1th back forward derivate
}
