% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_GEE.R
\name{gee_data.table}
\alias{gee_data.table}
\title{tidy_gee_json}
\usage{
gee_data.table(x, varnames = c("Lai", "Fpar", "FparLai_QC"))
}
\description{
tidy_gee_json
}
\examples{
\dontrun{
l = read_json("phenoflux212_MCD15A3H_006_463m_buffer.geojson")$features
x = l[[1]]
tidy_gee_json(x)
}
}
