% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_phenofit.R
\name{get_fitting2}
\alias{get_fitting2}
\title{get_fitting2}
\usage{
get_fitting2(l)
}
\arguments{
\item{l}{list with the element of c(\code{brks}, \code{fit}, \code{pheno})}
}
\description{
get_fitting2
}
