% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_func.R
\name{ggplot_1var}
\alias{ggplot_1var}
\title{global variables:
smooth_formula, span}
\usage{
ggplot_1var(
  x,
  varname = "APAR",
  color = "red",
  alpha = 0.3,
  span = 0.7,
  lwd = 1,
  ylab = NULL,
  labels = waiver()
)
}
\description{
global variables:
smooth_formula, span
}
