% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_async.R
\name{ggplot_multiAxis}
\alias{ggplot_multiAxis}
\title{ggplot_multiAxis}
\usage{
ggplot_multiAxis(p1, p2, show = TRUE)
}
\arguments{
\item{p1, p2}{should have axis.tick.y.right and axis.title.y.left}
}
\description{
ggplot_multiAxis
}
