% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPP_LUE_model.R
\name{LUE}
\alias{LUE}
\alias{LUE_Tscalar}
\alias{LUE_Wscalar}
\title{Light use efficiency GPP model}
\usage{
LUE_Tscalar(T, IGBP)

LUE_Wscalar(df, pheno_T)
}
\arguments{
\item{pheno_T}{data.frame with the columsn of \verb{site, year2, sos_date, eos_date}.}

\item{d_mod09a1}{data.frame with the columns of \verb{site, date, year2, LSWI, QC_flag}}
}
\description{
\code{LUE_Tscalar}: Air temperature constrain

\code{LUE_Wscalar}: Water constrain
}
