% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pheno_thermal.R
\name{n_continue}
\alias{n_continue}
\title{n_continue}
\usage{
n_continue(x, trs = 0, n = 3, type = "sos")
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Find the position of first n continous true \code{con}.}

\item{type}{'sos' means the left I (spring case),
'eos' means the right I (autumn case)}
}
\description{
n_continue
}
