% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pheno_thermal.R
\name{Pheno_thermal}
\alias{Pheno_thermal}
\title{Pheno_thermal}
\usage{
Pheno_thermal(x, t, trs = c(5, 10), n = 3, adjust = TRUE)
}
\arguments{
\item{x}{Numeric vector, temperature}

\item{t}{Date vector, corresponding date of \code{x}}

\item{n}{Integer, continous days of T >= trs\link{1} or T <= trs\link{2}}

\item{adjust}{If true:
\enumerate{
\item high latitude boreal regions will be fixed. If \code{all(T <= trs[1])},
growing season is set to Jun-Aug.
\item sos >= eos will be fixed, sos = pop-1; eos = pop+1.
}}
}
\description{
Thermal phenology based on temperature data.
}
\examples{
x = rnorm(365)
t = seq(as.Date("2010-01-01"), as.Date("2010-12-31"), by = "day")
}
