% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoAsync-package.R
\docType{package}
\name{PhenoAsync-package}
\alias{PhenoAsync}
\alias{PhenoAsync-package}
\title{PhenoAsync: Extract Remote Sensing Vegetation Phenology}
\description{
Async of vegetation carbon phenology and vegetation greennesss phenology.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
