% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenofit.R
\name{phenofit_plot}
\alias{phenofit_plot}
\title{plot_phenofit}
\usage{
phenofit_plot(
  obj,
  type = "all",
  methods,
  title = NULL,
  title.ylab = "Vegetation Index",
  IsPlot = TRUE,
  show.legend = TRUE,
  newpage = TRUE,
  ...
)
}
\arguments{
\item{obj}{\code{phenofit} object, list(INPUT, fit, seasons)}

\item{type}{one of c("season", "fitting", "pheno", "all")}

\item{methods}{Fine curve fitting methods, can be one or more of
\code{c('AG', 'Beck', 'Elmore', 'Zhang', 'Gu', 'Klos')}. Note that 'Gu' and 'Klos'
are very slow.
f not specified, it will be determined by phenofit options \code{methods_fine}.}

\item{title}{String, title of figure.}

\item{title.ylab}{String, title of \code{xlab} and \code{ylab}.}

\item{IsPlot}{boolean. If false, a ggplot object will be returned.}

\item{show.legend}{If now show legend, ggplot object will be returned, else
grid object will be returned.}

\item{newpage}{boolean, whether draw figure in a new page?}

\item{...}{other parameters to \code{\link[phenofit:plot_curvefits]{phenofit::plot_curvefits()}}}
}
\description{
plot_phenofit
}
\keyword{internel}
