% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_async.R
\name{scaleGPP_ByCoef}
\alias{scaleGPP_ByCoef}
\title{scaleGPP_ByCoef
scaled GPP into same range of coef, by IGBP}
\usage{
scaleGPP_ByCoef(d)
}
\arguments{
\item{d}{A data.frame with the columns of 'min' and 'max' and 'GPP'}
}
\description{
scaleGPP_ByCoef
scaled GPP into same range of coef, by IGBP
}
