% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_perc}
\alias{sum_perc}
\alias{mean_perc}
\title{sum_perc}
\usage{
sum_perc(x, perc = 0.8)

mean_perc(x, perc = 0.8)
}
\arguments{
\item{x}{Numeric vector}

\item{perc}{If the percentage of valid (good-quality) values less than \code{perc},
\code{NA} returned.}
}
\description{
sum_perc
}
\note{
\code{mean_perc} doesn't include the number of invalid values.
}
\examples{
x = c(1:8, NA, NA)
mean_perc(x)
}
