% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_GEE.R
\name{writeOGR_3by3}
\alias{writeOGR_3by3}
\title{writeOGR_3by3}
\usage{
writeOGR_3by3(st, scale = 500, prefix = "st212", outdir = "INPUT/shp")
}
\arguments{
\item{st}{data.table with the columns at least of \code{site}, \code{lon}, and \code{lat}.}

\item{scale}{in the unit of \code{m}}
}
\description{
writeOGR_3by3
}
\examples{
\dontrun{
st = st_212[, .(site, lon, lat, IGBP)]
writeOGR_3by3(st, 500, "agripheno")
}
}
