% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_lag.R
\name{corr_lag}
\alias{corr_lag}
\alias{corr_lags}
\alias{find_x_lag}
\title{Lag correlation between two variables}
\usage{
corr_lag(
  x,
  y = x,
  lag = 1,
  index = NULL,
  dates = NULL,
  doy_lims,
  verbose = FALSE
)

corr_lags(x, y, lags = -5:5, index, dates = NULL, doy_lims, ...)

find_x_lag(x, lag, index)
}
\arguments{
\item{x, y}{A numeric vector}

\item{lag}{The step of lag time}

\item{index}{A numeric vector, the subscript of x, e.g. \code{ydn}}

\item{dates}{Dates vector, dates of x}

\item{doy_lims}{constrain index in the range of \code{period}}

\item{...}{other parameters to \code{\link[=corr_lag]{corr_lag()}}}
}
\value{
\itemize{
\item \code{cor}: pearson correlation of lag step
\item \code{lag}:
}
\itemize{
\item positive: y is earlier;
\item negative: x is earlier.
}
}
\description{
Lag correlation between two variables
}
\note{
correlation needs to be adjusted because the length of observation is different
}
\examples{
# examples
x = sin(1:365)
y = x
index = 1:365
dates = seq(as.Date('2010-01-01'), as.Date('2010-12-31'), 'day')
corr_lags(x, y, index, dates, lag = 1)

}
