% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_seasons.R
\name{divide_seasons}
\alias{divide_seasons}
\title{divide_seasons}
\usage{
divide_seasons(d, sp, nptperyear = 23, lambda = 50, .movmean = TRUE)
}
\arguments{
\item{d}{data.frame of vegetation index}

\item{sp}{one line data.frame}

\item{.movmean}{moving mean by phenofit::wSG}

\item{sitename}{character}
}
\description{
divide_seasons
}
\note{
site-year may be not continuous.
}
