% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_seasons.R
\name{divide_seasons}
\alias{divide_seasons}
\title{divide_seasons}
\usage{
divide_seasons(
  d,
  nptperyear = 23,
  iters = 2,
  south = FALSE,
  igbp = "GRA",
  lambda = 100,
  wFUN = wTSM,
  r_max = 0.2,
  r_min = 0,
  r_minPeakHeight = 0.05,
  calendarYear = FALSE,
  .movmean = TRUE,
  .v_curve = FALSE,
  ...
)
}
\arguments{
\item{d}{data.frame, with the columns of \code{t}, \code{y} and \code{w}.}

\item{.movmean}{moving mean by phenofit::wSG}

\item{.v_curve}{If true, it will use V-curve theory to optimize Whittaker
parameter, lambda.}

\item{sp}{one line data.frame}

\item{sitename}{character}
}
\description{
divide_seasons
}
\note{
site-year may be not continuous.
}
